gusto_pos/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
alembic
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
uuid
